requirements.txt
fastapi
uvicorn
python-dotenv
openai
supabase-py
twilio
